+++ patches-core/src/modules/module_descriptor.rs
@@ -30,7 +30,7 @@ use super::parameter_map::ParameterValue;
 
 #[derive(Debug, Clone)]
 pub enum ParameterKind {
-    Float { min: f64, max: f64, default: f64 },
+    Float { min: f32, max: f32, default: f32 },
     Int   { min: i64, max: i64, default: i64 },
     Bool  { default: bool },
     Enum  { variants: &'static [&'static str], default: &'static str },
